+++ com.unity.netcode.gameobjects/Runtime/Core/NetworkBehaviour.cs
@@ -917,7 +917,7 @@ namespace Unity.Netcode
         }
 
         /// <summary>
-        /// In client-server contexts, this method is invoked on both the server and the local client of the owner when <see cref="Netcode.NetworkObject"/> ownership is assigned.
+        /// In client-server contexts, this method is invoked on both the server and the local client of the owner when <see cref="NetworkObject"/> ownership is assigned.
         /// In distributed authority contexts, this method is invoked on all clients connected to the session.
         /// </summary>
         public virtual void OnGainedOwnership() { }
@@ -952,7 +952,7 @@ namespace Unity.Netcode
         }
 
         /// <summary>
-        /// In client-server contexts, this method is invoked on the local client when it loses ownership of the associated <see cref="Netcode.NetworkObject"/>
+        /// In client-server contexts, this method is invoked on the local client when it loses ownership of the associated <see cref="NetworkObject"/>
         /// and on the server when any client loses ownership.
         /// In distributed authority contexts, this method is invoked on all clients connected to the session.
         /// </summary>